feat: update to 1.21.2-1.21.3 protocol i tried compiling for 1.21.4, and it seemed to work, but most clones' viaversion is not that updated yet so i just stayed at 1.21.3

+++ src/main/java/me/chayapak1/chomens_bot/plugins/SelfCarePlugin.java
@@ -1,5 +1,9 @@
 package me.chayapak1.chomens_bot.plugins;
 
+import me.chayapak1.chomens_bot.Bot;
+import me.chayapak1.chomens_bot.Configuration;
+import net.kyori.adventure.text.Component;
+import org.cloudburstmc.math.vector.Vector3d;
 import org.geysermc.mcprotocollib.network.Session;
 import org.geysermc.mcprotocollib.network.event.session.DisconnectedEvent;
 import org.geysermc.mcprotocollib.network.packet.Packet;
@@ -14,10 +18,6 @@ import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.inventory.C
 import org.geysermc.mcprotocollib.protocol.packet.ingame.clientbound.level.ClientboundGameEventPacket;
 import org.geysermc.mcprotocollib.protocol.packet.ingame.serverbound.ServerboundClientCommandPacket;
 import org.geysermc.mcprotocollib.protocol.packet.ingame.serverbound.inventory.ServerboundContainerClosePacket;
-import me.chayapak1.chomens_bot.Bot;
-import me.chayapak1.chomens_bot.Configuration;
-import net.kyori.adventure.text.Component;
-import org.cloudburstmc.math.vector.Vector3i;
 
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
@@ -81,7 +81,7 @@ public class SelfCarePlugin extends Bot.Listener {
 
         bot.position.addListener(new PositionPlugin.Listener() {
             @Override
-            public void positionChange(Vector3i position) {
+            public void positionChange(Vector3d position) {
                 SelfCarePlugin.this.positionChange();
             }
         });
